Added hashtags for direct FAQ linking
