+++ components/try.tsx
@@ -1,5 +1,4 @@
 import { Button } from "@/components/ui/button";
-import { ExternalLink } from "lucide-react";
 import Link from "next/link";
 
 export default function TryPear() {
