+++ types/faqItems.ts
@@ -1,4 +1,5 @@
 export interface FAQItem {
+  id: string;
   question: string;
   answer: JSX.Element;
 }
